Fix clippy hint

+++ dispatcher/src/sftp_downloader.rs
@@ -365,7 +365,7 @@ where
             .inc();
         metrics::BYTES_DOWNLOADED_COUNTER_VEC
             .with_label_values(&[&self.sftp_source.name])
-            .inc_by(bytes_copied as u64);
+            .inc_by(bytes_copied);
 
         if msg.remove {
             let unlink_result = sftp.unlink(remote_path);
